requirements.txt
crewai
crewai-tools
unstructured
python-dotenv
streamlit
serpapi
google-search-results
psycopg2-binary
azure-storage-blob
azure-identity
